Move BlockLocator tests
